+++ NewBuildOland/Entities/BlackWarrior.h
@@ -15,7 +15,7 @@ protected:
     std::vector<IntRect> getEastWalkAnim() override;
     std::vector<IntRect> getWestWalkAnim() override;
 
-    void updateMovement(double dt) override;
+    void updateMovement(float dt) override;
 
     virtual BlackWarrior* clone() const { return new BlackWarrior(*this); };
