+++ NewBuildOland/Entities/OtherPlayer.h
@@ -31,7 +31,7 @@ class OtherPlayer : public LivingEntity
         std::vector<IntRect> getEastWalkAnim() override;
         std::vector<IntRect> getWestWalkAnim() override;
 
-        void updateMovement(double dt) override;
+        void updateMovement(float dt) override;
 
         sf::Text nameText; //Display name
